[ref] playTime 분리
play time function class로  분리
