+++ README.md
@@ -16,5 +16,3 @@
 
 ## 💼 회의록
 
-추가 예정
-
